+++ src/com/intellij/idea/plugin/hybris/toolwindow/loggers/table/LoggersTable.kt
@@ -19,7 +19,9 @@
 package com.intellij.idea.plugin.hybris.toolwindow.loggers.table
 
 import com.intellij.idea.plugin.hybris.settings.RemoteConnectionSettings
+import com.intellij.idea.plugin.hybris.tools.logging.CxLoggerAccess
 import com.intellij.idea.plugin.hybris.tools.logging.CxLoggerModel
+import com.intellij.idea.plugin.hybris.tools.logging.CxLoggersState
 import com.intellij.openapi.project.Project
 import com.intellij.ui.ColoredTableCellRenderer
 import com.intellij.ui.JBColor
@@ -77,7 +79,7 @@ class LoggersTable : TableView<List<String>> {
         private const val serialVersionUID: Long = -1210838431719623644L
 
         fun of(project: Project, loggers: Map<String, CxLoggerModel>, connectionSettings: RemoteConnectionSettings): TableView<List<String>> {
-            val customCellRenderer = CustomCellRenderer(project, connectionSettings)
+            val customCellRenderer = CustomCellRenderer(CxLoggerAccess.getInstance(project).loggers(connectionSettings))
             val listTableModel = ListTableModel<List<String>>(
                 LoggerColumnInfo("Effective Level", COLUMN_LEVEL, customCellRenderer),
                 LoggerColumnInfo("Logger", COLUMN_LOGGER, customCellRenderer)
@@ -95,7 +97,7 @@ class LoggersTable : TableView<List<String>> {
     }
 }
 
-private class CustomCellRenderer(val project: Project, val connectionSettings: RemoteConnectionSettings) : ColoredTableCellRenderer() {
+private class CustomCellRenderer(private val loggersState: CxLoggersState) : ColoredTableCellRenderer() {
 
     override fun setToolTipText(text: String?) = Unit
 
@@ -104,7 +106,7 @@ private class CustomCellRenderer(val project: Project, val connectionSettings: R
 
         if (column == COLUMN_LOGGER) {
             append(stringValue, SimpleTextAttributes.GRAY_ATTRIBUTES)
-            //icon = CxLoggerAccess.getInstance(project).loggers(connectionSettings).get(stringValue).icon
+            icon = loggersState.get(stringValue).icon
             foreground = RenderingUtil.getForeground(table, selected)
             background = RenderingUtil.getBackground(table, selected)
             alignmentX = RIGHT_ALIGNMENT